+++ Changelog.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.1.0]
+### Added
+- Gradle files
+- Javadoc dependancy
+- License management plugin
+- JavaFX plugin
+- Gson dependency
+- UI for application
+- Ability to add courses
+- Ability to display all courses
+- Ability to display courses by department
+- Ability to save and load courses
+
+### Changed
+- Updated github-pages to reflect new documentation
+
 ## [v0.0.1] - 2022-02-22
 ### Added
 - Changelog.md to project root
@@ -16,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - .gitignore to the one provided
 - README.md to reflect the project
 
-[Unreleased]: https://github.com/TypeToSearch/cs2263_hw03/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/TypeToSearch/cs2263_hw03/compare/v0.1.0...HEAD
+[v0.1.0]: https://github.com/TypeToSearch/cs2263_hw03/releases/tag/v0.1.0
 [v0.0.1]: https://github.com/TypeToSearch/cs2263_hw03/releases/tag/v0.0.1